-관리자와 커뮤니티 모델 나누고 불필요한 코드 정리 1

+++ app/Http/Controllers/Admin/ContentsController.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace App\Http\Controllers\Content;
+namespace App\Http\Controllers\Admin;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use App\Content;
+use App\Admin\Content;
 
-class ContentController extends Controller
+class ContentsController extends Controller
 {
     public $content;
 
@@ -14,6 +14,7 @@ class ContentController extends Controller
     {
         $this->content = $content;
     }
+
     /**
      * Display a listing of the resource.
      *
@@ -21,24 +22,15 @@ class ContentController extends Controller
      */
     public function index()
     {
+        if (auth()->user()->cant('index', $this->content)) {
+            abort(403, '내용관리 목록 보기에 대한 권한이 없습니다.');
+        }
+
         $contents = $this->content->getContentList();
 
-        return view('content.index', [
+        return view('admin.content.index', [
             'contents' => $contents
         ]);
-    }
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        $params = $this->content->getContentView($id);
-        $skin = $params['content']->skin ? : 'default';
-
-        return view('content.'. $skin. '.show', $params);
     }
 
     /**
@@ -48,9 +40,13 @@ class ContentController extends Controller
      */
     public function create()
     {
+        if (auth()->user()->cant('create', Content::class)) {
+            abort(403, '내용관리 추가에 대한 권한이 없습니다.');
+        }
+
         $params = $this->content->getContentCreate();
 
-        return view('content.form', $params);
+        return view('admin.content.form', $params);
     }
 
     /**
@@ -61,35 +57,18 @@ class ContentController extends Controller
      */
     public function store(Request $request)
     {
+        if (auth()->user()->cant('create', Content::class)) {
+            abort(403, '내용관리 추가에 대한 권한이 없습니다.');
+        }
+
         $rules = $this->rules();
         $rules = array_add($rules, 'content_id', 'bail|required|max:20|unique:contents|regex:/^[a-zA-Z0-9_]+$/');
 
         $this->validate($request, $rules, $this->messages());
 
         $result = $this->content->storeContent($request);
-        
+
         return redirect(route('contents.edit', $result));
-    }
-
-    public function rules()
-    {
-        return [
-            'subject' => 'bail|required',
-            'content' => 'bail|required',
-        ];
-    }
-
-    // 에러 메세지
-    public function messages()
-    {
-        return [
-            'content_id.required' => 'ID를 입력해 주세요.',
-            'content_id.max' => 'ID는 20자리를 넘길 수 없습니다.',
-            'content_id.unique' => '이미 등록된 ID입니다. 다른 ID를 입력해 주세요.',
-            'content_id.regex' => 'ID는 20자 이내의 영문자, 숫자, _ 만 가능합니다.',
-            'subject.required'  => '제목을 입력해 주세요.',
-            'content.required'  => '내용을 입력해 주세요.',
-        ];
     }
 
     /**
@@ -100,9 +79,13 @@ class ContentController extends Controller
      */
     public function edit($id)
     {
+        if (auth()->user()->cant('update', $this->content)) {
+            abort(403, '내용관리 수정에 대한 권한이 없습니다.');
+        }
+
         $params = $this->content->getContentEdit($id);
 
-        return view('content.form', $params);
+        return view('admin.content.form', $params);
     }
 
     /**
@@ -114,6 +97,10 @@ class ContentController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (auth()->user()->cant('update', $this->content)) {
+            abort(403, '내용관리 수정에 대한 권한이 없습니다.');
+        }
+
         $this->validate($request, $this->rules(), $this->messages());
 
         $result = $this->content->updateContent($request, $id);
@@ -135,8 +122,33 @@ class ContentController extends Controller
      */
     public function destroy($id)
     {
+        if (auth()->user()->cant('delete', $this->content)) {
+            abort(403, '내용관리 삭제에 대한 권한이 없습니다.');
+        }
+
         $this->content->deleteContent($id);
 
         return redirect()->back();
+    }
+
+    public function rules()
+    {
+        return [
+            'subject' => 'bail|required',
+            'content' => 'bail|required',
+        ];
+    }
+
+    // 에러 메세지
+    public function messages()
+    {
+        return [
+            'content_id.required' => 'ID를 입력해 주세요.',
+            'content_id.max' => 'ID는 20자리를 넘길 수 없습니다.',
+            'content_id.unique' => '이미 등록된 ID입니다. 다른 ID를 입력해 주세요.',
+            'content_id.regex' => 'ID는 20자 이내의 영문자, 숫자, _ 만 가능합니다.',
+            'subject.required'  => '제목을 입력해 주세요.',
+            'content.required'  => '내용을 입력해 주세요.',
+        ];
     }
 }
